RAG with Ollama and DB/requirements.txt
flask
flask-session
langchain
langchain-community
langchain-core
langchain-ollama
python-dotenv
psycopg2-binary
pgvector
pdfplumber
faiss-cpu
redis
